+++ lineparser.lua
@@ -1,4 +1,4 @@
-package.cpath = 'externals/?/lib?.dylib;expernals/?/?.so;' .. package.cpath
+package.cpath = 'externals/?/lib?.dylib;externals/?/lib?.so;' .. package.cpath
 
 local lpeg = require 'lpeg'
 local write = io.write