+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-**Most recent upgrades**
+<!--**Most recent upgrades**-->
+## 0.6
 
-. ~~Snap BPM change to the next playing phrase (and not the current one).~~
+. `ga`command generates random amplitudes for each and every note in the playing phrase
+
+, `rl`, `rr` commands rotate, one/various positions, left or right, respectively, each note(s)/cc
+
+. Upgrade notes/cc sync process
+
+## 0.5.37
 
 . New meaning to `/` command. Eg. `/8`, phrase duration will be 8 times longer.